+++ index.js
@@ -1,7 +1,7 @@
 const instance_skel = require('../../instance_skel');
 const http          = require('http');
 const request       = require('request');
-let debug, log;
+var debug, log;
 
 /**
  * Companion instance class for the Epiphan Pearl.
@@ -40,8 +40,8 @@ class EpiphanPearl extends instance_skel {
 		this.CHOICES_RECORDERS = [];
 
 		this.CHOICES_STARTSTOP = [
-			{id: 1, label: 'Start'},
-			{id: 0, label: 'Stop'}
+			{id: 1, label: 'Start', action: 'start'},
+			{id: 0, label: 'Stop', action: 'stop'}
 		];
 
 
@@ -56,6 +56,13 @@ class EpiphanPearl extends instance_skel {
 	 * @since 1.0.0
 	 */
 	actions(system) {
+		const startStopOption = {
+			type: 'dropdown',
+			id: 'startStopAction',
+			label: 'Action',
+			choices: this.CHOICES_STARTSTOP
+		};
+
 		this.setActions({
 			'channelChangeLayout': {
 				label: 'Change channel layout',
@@ -66,25 +73,8 @@ class EpiphanPearl extends instance_skel {
 					choices: this.CHOICES_CHANNELS_LAYOUTS
 				}]
 			},
-			'channelRecording': {
-				label: 'Channel start/stop recording',
-				options: [
-					{
-						type: 'dropdown',
-						id: 'channelId',
-						label: 'Channel',
-						choices: this.CHOICES_CHANNELS
-					},
-					{
-						type: 'dropdown',
-						id: 'channelAction',
-						label: 'Action',
-						choices: this.CHOICES_STARTSTOP
-					},
-				],
-			},
 			'channelStreaming': {
-				label: 'Channel start/stop streaming',
+				label: 'Start/stop streaming',
 				options: [
 					{
 						type: 'dropdown',
@@ -93,16 +83,11 @@ class EpiphanPearl extends instance_skel {
 						choices: this.CHOICES_CHANNELS_PUBLISHERS,
 						tooltip: 'If a channel has only one "publisher" or "stream" then you jst select all. Else you can pick the "publisher" you want to start/stop'
 					},
-					{
-						type: 'dropdown',
-						id: 'channelAction',
-						label: 'Action',
-						choices: this.CHOICES_STARTSTOP
-					},
+					startStopOption
 				],
 			},
 			'recorderRecording': {
-				label: 'Recorder start/stop recording',
+				label: 'Start/stop recording',
 				options: [
 					{
 						type: 'dropdown',
@@ -110,12 +95,7 @@ class EpiphanPearl extends instance_skel {
 						label: 'Recorder',
 						choices: this.CHOICES_RECORDERS
 					},
-					{
-						type: 'dropdown',
-						id: 'recorderAction',
-						label: 'Action',
-						choices: this.CHOICES_STARTSTOP
-					},
+					startStopOption
 				],
 			},
 		});
@@ -130,34 +110,32 @@ class EpiphanPearl extends instance_skel {
 	 */
 	action(action) {
 		let type = 'get', url, body = {};
-		let channelId, layoutId, channelAction, publishersId;
+		let channelId, layoutId, optionAction, publishersId, startStopAction;
 
 		switch (action.action) {
 			case 'channelChangeLayout':
 				[channelId, layoutId] = action.options.channelIdlayoutId.split('-');
-				type                  = 'put';
-				url                   = '/api/channels/' + channelId + '/layouts/active';
-				body                  = JSON.stringify({id: layoutId});
+
+				type = 'put';
+				url  = '/api/channels/' + channelId + '/layouts/active';
+				body = {id: layoutId};
 				break;
 			case 'channelStreaming':
 				[channelId, publishersId] = action.options.channelIdpublisherId.split('-');
-				channelAction             = this.CHOICES_STARTSTOP.find(function (e) { return e.id === action.options.channelAction; });
-				type                      = 'post';
+				startStopAction           = this._getStartStopActionFromOptions(action.options);
 
-				if (publishersId !== 'all') {
-					url = '/api/channels/' + channelId + '/publishers/' + publishersId + '/control/' + channelAction;
-				} else {
-					url = '/api/channels/' + channelId + '/publishers/control/' + channelAction;
-				}
-
-				break;
-			case 'channelRecording':
 				type = 'post';
-				//url  = '/api/channels/' + action.options.channelId + '/control/start';
+				if (publishersId !== 'all') {
+					url = '/api/channels/' + channelId + '/publishers/' + publishersId + '/control/' + startStopAction;
+				} else {
+					url = '/api/channels/' + channelId + '/publishers/control/' + startStopAction;
+				}
 				break;
 			case 'recorderRecording':
+				startStopAction = this._getStartStopActionFromOptions(action.options);
+
 				type = 'post';
-				//url  = '/api/recorders/' + action.options.channelId + '/control/start';
+				url  = '/api/recorders/' + action.options.recorderId + '/control/' + startStopAction;
 				break;
 			default:
 				return;
@@ -277,6 +255,29 @@ class EpiphanPearl extends instance_skel {
 	}
 
 	/**
+	 * INTERNAL: handler of status changes
+	 *
+	 * @private
+	 * @since 1.0.0
+	 */
+	_setStatus(level, message = '') {
+		this.debug('debug', '_set_status(level, message):(' + level + ',' + message + ')');
+		this.status(level, message);
+	}
+
+	/**
+	 * INTERNAL: Get action from the options for start and stop
+	 *
+	 * @private
+	 * @since 1.0.0
+	 */
+	_getStartStopActionFromOptions(options) {
+		return this.CHOICES_STARTSTOP.find((e) => {
+			return e.id === parseInt(options.startStopAction);
+		}).action;
+	}
+
+	/**
 	 * INTERNAL: Setup and send request
 	 *
 	 * @param {String} type - post, get, put
@@ -287,35 +288,40 @@ class EpiphanPearl extends instance_skel {
 	 * @private
 	 * @since 1.0.0
 	 */
-	_sendRequest(type, url, body, callback = null) {
+	_sendRequest(type, url, body = {}, callback = () => {
+	}) {
 		let self      = this;
 		const apiHost = this.config.host,
 			  baseUrl = 'http://' + apiHost;
 
 		if (url === null || url === '') {
+			this._setStatus(this.STATUS_ERROR, 'No URL given for _sendRequest');
+			this.log('error', 'No URL given for _sendRequest');
 			return false;
 		}
 
 		if (this.defaultRequest === undefined) {
-			this.status(this.STATUS_ERROR, 'Error, no request interface...');
+			this._setStatus(this.STATUS_ERROR, 'Error, no request interface...');
 			this.log('error', 'Error, no request interface...');
 			return false;
 		}
 
 		type = type.toUpperCase();
 		if (['GET', 'POST', 'PUT'].indexOf(type) === -1) {
-			this.status(this.STATUS_ERROR, 'Wrong request type: ' + type);
+			this._setStatus(this.STATUS_ERROR, 'Wrong request type: ' + type);
 			this.log('error', 'Wrong request type: ' + type);
 			return false;
 		}
 
+		// Check if body is not empty
 		if (body === undefined) {
 			body = {};
 		}
 
-		if (callback === undefined || callback == null) {
+		// Check if we have a valid callback
+		if (callback === undefined || callback == null || typeof callback !== 'function') {
 			callback = () => {
-			}
+			};
 		}
 
 		let requestUrl = baseUrl + url;
@@ -326,27 +332,27 @@ class EpiphanPearl extends instance_skel {
 				uri: requestUrl,
 				json: body
 			},
-			function (error, response, body) {
+			(error, response, body) => {
 				self.debug('info', JSON.stringify(error));
 				//self.debug('info', JSON.stringify(response));
 				self.debug('info', JSON.stringify(body));
 
 				if (error && error.code === 'ETIMEDOUT') {
-					self.status(self.STATUS_ERROR);
+					self._setStatus(self.STATUS_ERROR, 'Connection timeout while connecting to ' + requestUrl);
 					self.log('error', 'Connection timeout while connecting to ' + requestUrl);
 					callback(error);
 					return;
 				}
 
 				if (error && error.code === 'ECONNREFUSED') {
-					self.status(self.STATUS_ERROR);
+					self._setStatus(self.STATUS_ERROR, 'Connection refused for ' + requestUrl);
 					self.log('error', 'Connection refused for ' + requestUrl);
 					callback(error);
 					return;
 				}
 
 				if (error && error.connect === true) {
-					self.status(self.STATUS_ERROR);
+					self._setStatus(self.STATUS_ERROR, 'Read timeout waiting for response from: ' + requestUrl);
 					self.log('error', 'Read timeout waiting for response from: ' + requestUrl);
 					callback(error);
 					return;
@@ -354,14 +360,14 @@ class EpiphanPearl extends instance_skel {
 
 				if (response &&
 					(response.statusCode < 200 || response.statusCode > 299)) {
-					self.status(self.STATUS_ERROR);
+					self._setStatus(self.STATUS_ERROR, 'Non-successful response status code: ' + http.STATUS_CODES[response.statusCode] + ' ' + requestUrl);
 					self.log('error', 'Non-successful response status code: ' + http.STATUS_CODES[response.statusCode] + ' ' + requestUrl);
 					callback(error);
 					return;
 				}
 
 				if (body.status && body.status !== 'ok') {
-					self.status(self.STATUS_ERROR);
+					self._setStatus(self.STATUS_ERROR, 'Non-successful response from pearl: ' + requestUrl + ' - ' + (body.message ? body.message : 'No error message'));
 					self.log('error', 'Non-successful response from pearl: ' + requestUrl + ' - ' + (body.message ? body.message : 'No error message'));
 					callback(error);
 					return;
@@ -372,7 +378,7 @@ class EpiphanPearl extends instance_skel {
 					result = body.result;
 				}
 
-				self.status(this.STATUS_OK);
+				self._setStatus(self.STATUS_OK);
 				callback(null, result);
 			});
 	}
@@ -491,14 +497,14 @@ class EpiphanPearl extends instance_skel {
 		let self = this;
 
 		// Get all channels available
-		let temp_channel = [];
+		let temp_channels = [];
 		this._sendRequest('get', '/api/channels', {}, (err, channels) => {
 			if (err) {
 				return;
 			}
 			for (let a in channels) {
 				let channel = channels[a];
-				temp_channel.push({
+				temp_channels.push({
 					id: channel.id,
 					label: channel.name,
 					layouts: [],
@@ -509,13 +515,13 @@ class EpiphanPearl extends instance_skel {
 
 			self.debug('info', 'Updating CHOICES_CHANNELS and call actions()');
 			// Update the master channel selector
-			self.CHOICES_CHANNELS = temp_channel.slice();
+			self.CHOICES_CHANNELS = temp_channels.slice();
 			// Update dropdowns
 			self.actions(system);
 		});
 
 		// For every channel get layouts and populate/update actions()
-		let temp_layout = [];
+		let temp_layouts = [];
 		for (let b in this.CHOICES_CHANNELS) {
 			let channel = this.CHOICES_CHANNELS[b];
 
@@ -525,7 +531,7 @@ class EpiphanPearl extends instance_skel {
 				}
 				for (let c in layouts) {
 					let layout = layouts[c];
-					temp_layout.push({
+					temp_layouts.push({
 						id: channel.id + '-' + layout.id,
 						label: channel.label + ' - ' + layout.name
 					});
@@ -533,7 +539,7 @@ class EpiphanPearl extends instance_skel {
 
 				self.debug('info', 'Updating CHANNEL_LAYOUTS and call actions()');
 				// Update the master channel selector
-				self.CHOICES_CHANNELS_LAYOUTS = temp_layout.slice();
+				self.CHOICES_CHANNELS_LAYOUTS = temp_layouts.slice();
 				// Update dropdowns
 				self.actions(system);
 			});
@@ -552,7 +558,7 @@ class EpiphanPearl extends instance_skel {
 				}
 
 				temp_publishers.push({
-					id: 'all',
+					id: channel.id + '-all',
 					label: channel.label + ' - All'
 				});
 				for (let b in channels[a].publishers) {
@@ -575,18 +581,26 @@ class EpiphanPearl extends instance_skel {
 			self.actions(system);
 		});
 
+		// Get all recorders
+		let temp_recoders = [];
+		this._sendRequest('get', '/api/recorders', {}, (err, recoders) => {
+			if (err) {
+				return;
+			}
+			for (let a in recoders) {
+				let recoder = recoders[a];
+				temp_recoders.push({
+					id: recoder.id,
+					label: recoder.name
+				});
+			}
 
-		// GET- /api/channels
-		// GET- /api/channels/?publishers=yes
-		// GET- /api/channels/status?publishers=yes&encoders=yes
-		// GET- /api/channels/<id>/status
-		// GET- /api/channels/<id>/layouts
-		// PUT- /api/channels/<id>/layouts/active - {"id":"3"}
-		// GET- /api/channels/<id>/publishers
-		// GET- /api/recorders
-		// GET- /api/recorders/status
-
-		//this._sendRequest();
+			self.debug('info', 'Updating CHOICES_RECORDERS and call actions()');
+			// Update the master channel selector
+			self.CHOICES_RECORDERS = temp_recoders.slice();
+			// Update dropdowns
+			self.actions(system);
+		});
 	}
 }